models update

+++ app/src/main/java/com/github/dawidhyzy/kotlinvsjava/jv/api/model/Snow.java
@@ -6,13 +6,13 @@ import com.google.gson.annotations.SerializedName;
 public class Snow {
 
     @SerializedName("3h")
-    private int _3h;
+    private Double _3h;
 
-    public int get3h() {
+    public Double get3h() {
         return _3h;
     }
 
-    public void set3h(int _3h) {
+    public void set3h(Double _3h) {
         this._3h = _3h;
     }
